+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Enhance `MixedNamesRule` with symbol table information.
+- Exclude class constructors in `ConstructorCreateRule`.
 
 ### Fixed